Fixed a bug in lreg benchmark

+++ test_suites/linear_regression/linear_regression_pthread.c
@@ -156,7 +156,7 @@ int main(int argc, char *argv[])
    /* Barrier, wait for all threads to finish */
    for (i = 0; i < num_threads; i++)
    {
-	  int ret_val;
+	  long long ret_val;
 	  CHECK_ERROR(pthread_join(tid_args[i].tid, (void **)(void*)&ret_val) != 0);
 	  CHECK_ERROR(ret_val != 0);
